backend/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
google-auth==2.3.3
gunicorn==20.0.4
idna==3.3
kubernetes==20.13.0
oauthlib==3.1.1
pip-autoremove==0.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
websocket-client==1.2.3
channels==3.0.4